Fix warning about unknown sizes for weak symbols provided by linker scripts
For example RTEMS has a weak symbol Tls_Alignment that is provided by the
linker script. As this is an ABS symbol we do not need a size for it since
at runtime only the address will be used. Using `&Tls_Alignment` will load
the untagged value from the cap_table and should just work so there is no
need to warn about a symbol with zero size here.